Splitting up specs
Our old spec file (kairos.spec.js) was very large (around 1500 lines)
so, I've split it up into 2 files, one for KairosTimeFrame and the
other for KairosCollection. The specs themselves were not changed,
though, the error message should be better now, as there will not
be a space between "Kairos" and "TimeFrame" or "Collection".